requirements.txt
feedparser>=6.0.0
requests>=2.31.0
pyyaml>=6.0
python-frontmatter>=1.1.0
arxiv>=2.1.0
tweepy>=4.14.0
python-dateutil>=2.8.0
